Mark this repo as EOL
We discovered that the GNOME 3.28 runtime currently contains GnuTLS
3.5.3, indicating there has been no security updates for over two years
now. (The latest stable version is 3.5.18.)

This is not an isolated problem, and fixing only GnuTLS is not a
solution. Epiphany cannot return to Flathub until proper security
support is provided by the entire runtime.

Epiphany Technology Preview remains supported, albeit hesitantly. There
are no doubt security problems there as well, but all the software in
the master runtime is more recent, so there should at least be fewer
such issues.