Include elementary stylesheet
Backport of https://gitlab.gnome.org/GNOME/epiphany/-/merge_requests/858

When running under Pantheon, we need the stylesheet to be available
inside the sandbox to avoid falling back to Adwaita. In the past
we relied on the Flatpak GTK extensions, but this isn't going to
work going forward for ElemetnaryOS 6 version of the stylesheet as
it provides multiple versions of the gtk theme, all with different
ids.